+++ Gruntfile.js
@@ -19,13 +19,11 @@ module.exports = function( grunt ) {
 
 
     // Load the NPM tasks.
-    grunt.loadNpmTasks( 'grunt-contrib-concat' )
     grunt.loadNpmTasks( 'grunt-contrib-watch' )
     grunt.loadNpmTasks( 'grunt-contrib-jshint' )
     grunt.loadNpmTasks( 'grunt-contrib-qunit' )
     grunt.loadNpmTasks( 'grunt-contrib-copy' )
     grunt.loadNpmTasks( 'grunt-contrib-less' )
-    grunt.loadNpmTasks( 'grunt-contrib-clean' )
     grunt.loadNpmTasks( 'grunt-contrib-cssmin' )
     grunt.loadNpmTasks( 'grunt-contrib-uglify' )
 
@@ -41,42 +39,39 @@ module.exports = function( grunt ) {
         // Set up the directories.
         dirs: {
             tests: 'tests',
-            src: {
-                raw: '_raw',
-                demos: '_raw/demo',
-                pickers: '_raw/lib',
-                themes: '_raw/lib/themes',
-                translations: '_raw/lib/translations'
+            lib: {
+                src: 'lib',
+                min: 'lib/compressed'
             },
-            dest: {
-                demos: 'demo',
-                pickers: 'lib',
-                themes: 'lib/themes',
-                translations: 'lib/translations'
+            themes: {
+                src: 'lib/themes-source',
+                dest: 'lib/themes',
+                min: 'lib/compressed/themes'
             },
-            min: {
-                pickers: 'lib/compressed',
-                themes: 'lib/compressed/themes',
-                translations: 'lib/compressed/translations'
-            }
-        },
-
-
-        // Clean the destination files and directories.
-        clean: {
-            demos: [ '<%= dirs.dest.demos %>', '*.htm' ],
-            pickers: [ '<%= dirs.dest.pickers %>/*.js' ],
-            themes: [ '<%= dirs.dest.themes %>' ],
-            translations: [ '<%= dirs.dest.translations %>' ],
-            pkg: [ '<%= dirs.dest.pickers %>', '<%= pkg.name %>.jquery.json', '*.md' ]
+            translations: {
+                src: 'lib/translations',
+                min: 'lib/compressed/translations'
+            },
+            docs: {
+                src: '_docs',
+                dest: '.'
+            },
+            demo: {
+                images: 'demo/images',
+                scripts: 'demo/scripts',
+                styles: {
+                    src: 'demo/styles/less',
+                    dest: 'demo/styles/css'
+                }
+            },
         },
 
 
         // Generate static HTML templates.
         htmlify: {
-            demos: {
+            docs: {
                 expand: true,
-                cwd: '<%= dirs.src.raw %>',
+                cwd: '<%= dirs.docs.src %>',
                 src: [ '/!(base|hero)*.htm' ],
                 dest: '',
                 base: '/base.htm'
@@ -86,18 +81,6 @@ module.exports = function( grunt ) {
 
         // Copy over files to destination directions.
         copy: {
-            demos: {
-                expand: true,
-                cwd: '<%= dirs.src.demos %>',
-                src: [ 'styles/.css', 'images/*.{png,ico}' ],
-                dest: '<%= dirs.dest.demos %>'
-            },
-            translations: {
-                expand: true,
-                cwd: '<%= dirs.src.translations %>',
-                src: [ '*' ],
-                dest: '<%= dirs.dest.translations %>'
-            },
             pkg: {
                 options: {
                     processContent: function( content ) {
@@ -107,10 +90,10 @@ module.exports = function( grunt ) {
                 files: [
                     { '<%= pkg.name %>.jquery.json': 'package.json' },
                     { 'bower.json': 'package.json' },
-                    { 'README.md': '<%= dirs.src.raw %>/README.md' },
-                    { 'LICENSE.md': '<%= dirs.src.raw %>/LICENSE.md' },
-                    { 'CHANGELOG.md': '<%= dirs.src.raw %>/CHANGELOG.md' },
-                    { 'CONTRIBUTING.md': '<%= dirs.src.raw %>/CONTRIBUTING.md' }
+                    { 'README.md': '<%= dirs.docs.src %>/README.md' },
+                    { 'LICENSE.md': '<%= dirs.docs.src %>/LICENSE.md' },
+                    { 'CHANGELOG.md': '<%= dirs.docs.src %>/CHANGELOG.md' },
+                    { 'CONTRIBUTING.md': '<%= dirs.docs.src %>/CONTRIBUTING.md' }
                 ]
             }
         },
@@ -121,40 +104,21 @@ module.exports = function( grunt ) {
             options: {
                 style: 'expanded'
             },
-            demos: {
+            demo: {
                 files: {
-                    '<%= dirs.dest.demos %>/styles/main.css': '<%= dirs.src.demos %>/styles/base.less'
+                    '<%= dirs.demo.styles.dest %>/main.css': '<%= dirs.demo.styles.src %>/base.less'
                 }
             },
             themes: {
                 files: {
-                    '<%= dirs.dest.themes %>/default.css': [ '<%= dirs.src.themes %>/base.less', '<%= dirs.src.themes %>/default.less' ],
-                    '<%= dirs.dest.themes %>/classic.css': [ '<%= dirs.src.themes %>/base.less', '<%= dirs.src.themes %>/classic.less' ],
-                    '<%= dirs.dest.themes %>/default.date.css': [ '<%= dirs.src.themes %>/base.date.less', '<%= dirs.src.themes %>/default.date.less' ],
-                    '<%= dirs.dest.themes %>/default.time.css': [ '<%= dirs.src.themes %>/base.time.less', '<%= dirs.src.themes %>/default.time.less' ],
-                    '<%= dirs.dest.themes %>/classic.date.css': [ '<%= dirs.src.themes %>/base.date.less', '<%= dirs.src.themes %>/classic.date.less' ],
-                    '<%= dirs.dest.themes %>/classic.time.css': [ '<%= dirs.src.themes %>/base.time.less', '<%= dirs.src.themes %>/classic.time.less' ],
-                    '<%= dirs.dest.themes %>/rtl.css': [ '<%= dirs.src.themes %>/rtl.less' ]
+                    '<%= dirs.themes.dest %>/default.css': [ '<%= dirs.themes.src %>/base.less', '<%= dirs.themes.src %>/default.less' ],
+                    '<%= dirs.themes.dest %>/classic.css': [ '<%= dirs.themes.src %>/base.less', '<%= dirs.themes.src %>/classic.less' ],
+                    '<%= dirs.themes.dest %>/default.date.css': [ '<%= dirs.themes.src %>/base.date.less', '<%= dirs.themes.src %>/default.date.less' ],
+                    '<%= dirs.themes.dest %>/default.time.css': [ '<%= dirs.themes.src %>/base.time.less', '<%= dirs.themes.src %>/default.time.less' ],
+                    '<%= dirs.themes.dest %>/classic.date.css': [ '<%= dirs.themes.src %>/base.date.less', '<%= dirs.themes.src %>/classic.date.less' ],
+                    '<%= dirs.themes.dest %>/classic.time.css': [ '<%= dirs.themes.src %>/base.time.less', '<%= dirs.themes.src %>/classic.time.less' ],
+                    '<%= dirs.themes.dest %>/rtl.css': [ '<%= dirs.themes.src %>/rtl.less' ]
                 }
-            }
-        },
-
-
-        // Concatenate the files and add the banner.
-        concat: {
-            options: {
-                process: function( content ) {
-                    return grunt.template.process( content, { delimiters: 'curly' } )
-                }
-            },
-            demos: {
-                files: { '<%= dirs.dest.demos %>/scripts/main.js': '<%= dirs.src.demos %>/scripts/*.js' }
-            },
-            pickers: {
-                expand: true,
-                cwd: '<%= dirs.src.pickers %>',
-                src: [ '*.js' ],
-                dest: '<%= dirs.dest.pickers %>'
             }
         },
 
@@ -165,14 +129,14 @@ module.exports = function( grunt ) {
                 jshintrc: true
             },
             gruntfile: 'Gruntfile.js',
-            demos: [ '<%= dirs.src.demos %>/scripts/base.js' ],
-            pickers: [
+            demo: [ '<%= dirs.demo.scripts %>/demo.js' ],
+            lib: [
                 '<%= dirs.tests %>/units/*.js',
-                '<%= dirs.dest.pickers %>/**/*.js',
+                '<%= dirs.lib.src %>/**/*.js',
 
                 // Ignore the legacy and minified files.
-                '!<%= dirs.dest.pickers %>/legacy.js',
-                '!<%= dirs.dest.pickers %>/compressed/**/*.js'
+                '!<%= dirs.lib.src %>/legacy.js',
+                '!<%= dirs.lib.src %>/compressed/**/*.js'
             ]
         },
 
@@ -182,64 +146,56 @@ module.exports = function( grunt ) {
             options: {
                 preserveComments: 'some'
             },
-            pickers: {
+            lib: {
                 files: [
                     {
                         expand : true,
-                        cwd : '<%= dirs.dest.pickers %>',
+                        cwd : '<%= dirs.lib.src %>',
                         src   : [ '**/*.js', '!compressed/**/*.js' ],
-                        dest : '<%= dirs.dest.pickers %>/compressed'
+                        dest : '<%= dirs.lib.min %>'
                     }
                 ]
             }
         },
         cssmin: {
-            pickers: {
+            lib: {
                 expand: true,
-                cwd: '<%= dirs.dest.pickers %>',
+                cwd: '<%= dirs.themes.dest %>',
                 src: [ '**/*.css', '!compressed/**/*.css' ],
-                dest: '<%= dirs.dest.pickers %>/compressed'
+                dest: '<%= dirs.themes.min %>'
             }
         },
 
 
         // Unit test the files.
         qunit: {
-            pickers: [ '<%= dirs.tests %>/units/all.htm' ]
+            lib: [ '<%= dirs.tests %>/units/all.htm' ]
         },
 
 
         // Watch the project files.
         watch: {
-            gruntfile: {
-                files: [ 'Gruntfile.js' ],
-                tasks: [ 'jshint:gruntfile', 'default' ]
-            },
             quick: {
                 files: [
-                    '<%= dirs.src.raw %>/*.htm',
-                    '<%= dirs.src.raw %>/*.md',
-                    '<%= dirs.src.demos %>/styles/*.less', '<%= dirs.src.demos %>/scripts/*.js',
-                    '<%= dirs.src.pickers %>/**/*.js', '<%= dirs.src.pickers %>/themes/*.css',
-                    '<%= dirs.src.themes %>/*.less',
-                    '<%= dirs.src.translations %>/*.js'
+                    '<%= dirs.docs.src %>/*.htm',
+                    '<%= dirs.docs.src %>/*.md',
+                    '<%= dirs.demo.src %>/styles/*.less',
+                    '<%= dirs.themes.src %>/*.less'
                 ],
                 tasks: [ 'quick' ]
             },
-            demos: {
+            docs: {
                 files: [
-                    '<%= dirs.src.raw %>/*.htm',
-                    '<%= dirs.src.demos %>/styles/*.less', '<%= dirs.src.demos %>/scripts/*.js'
+                    '<%= dirs.docs.src %>/*.htm',
+                    '<%= dirs.docs.src %>/*.md'
                 ],
-                tasks: [ 'demo' ]
+                tasks: [ 'docs' ]
             },
-            pickers: {
+            themes: {
                 files: [
-                    '<%= dirs.src.pickers %>/**/*.js', '<%= dirs.src.pickers %>/themes/*.css',
-                    '<%= dirs.src.themes %>/*.less',
-                    '<%= dirs.src.translations %>/*.js'
+                    '<%= dirs.themes.src %>/*.less'
                 ],
-                tasks: [ 'picker' ]
+                tasks: [ 'themes' ]
             }
         },
 
@@ -263,11 +219,12 @@ module.exports = function( grunt ) {
 
     // Register the tasks.
     // * `htmlify` and `copy:pkg` should come after `uglify` because some package files measure `.min` file sizes.
-    grunt.registerTask( 'default', [ 'clean', 'concat', 'copy:demos', 'copy:translations', 'less', 'jshint', 'qunit', 'uglify', 'cssmin', 'htmlify', 'copy:pkg' ] )
-    grunt.registerTask( 'quick', [ 'concat', 'copy:demos', 'copy:translations', 'less', 'uglify', 'cssmin', 'htmlify', 'copy:pkg' ] )
-    grunt.registerTask( 'picker', [ 'clean:pickers', 'concat:pickers', 'copy:translations', 'less:themes', 'jshint:pickers', 'qunit:pickers', 'uglify:pickers', 'cssmin:pickers' ] )
-    grunt.registerTask( 'demo', [ 'clean:demos', 'concat:demos', 'copy:demos', 'less:demos', 'jshint:demos', 'htmlify:demos' ] )
-    grunt.registerTask( 'travis', [ 'jshint:pickers', 'qunit:pickers' ] )
+    grunt.registerTask( 'default', [ 'less', 'jshint', 'qunit', 'uglify', 'cssmin', 'htmlify', 'copy:pkg' ] )
+    grunt.registerTask( 'quick', [ 'less', 'uglify', 'cssmin', 'htmlify', 'copy:pkg' ] )
+    grunt.registerTask( 'themes', [ 'less:themes' ] )
+    grunt.registerTask( 'demo', [ 'less:demo', 'jshint:demo' ] )
+    grunt.registerTask( 'docs', [ 'copy:pkg', 'htmlify:docs' ] )
+    grunt.registerTask( 'travis', [ 'jshint:lib', 'qunit:lib' ] )