+++ tests/units/time.js
@@ -148,6 +148,19 @@ test( '`interval`', function() {
     strictEqual( picker.get( 'interval' ), 120, 'Interval unaffected by non-integer' )
 })
 
+test( '`clear`', function() {
+
+    var picker = this.picker
+
+    strictEqual( picker.get('select'), null, 'Starts off without a selection' )
+
+    picker.set('select', new Date())
+    notStrictEqual( picker.get('select'), null, 'A selection has been added' )
+
+    picker.clear()
+    strictEqual( picker.get('select'), null, 'Clears out selection' )
+})
+
 test( '`select`', function() {
 
     var picker = this.picker
@@ -324,9 +337,9 @@ test( '`min` using JS dates', function() {
 
     // Using JavaScript date objects
     var dateObject = new Date()
-    dateObject.setHours(4,20)
+    dateObject.setHours(4,30)
     picker.set( 'min', dateObject )
-    strictEqual( picker.get( 'min' ).pick, 260, '`min` using a JS date: ' + picker.get( 'min', 'HH:i' ) )
+    strictEqual( picker.get( 'min' ).pick, 270, '`min` using a JS date: ' + picker.get( 'min', 'HH:i' ) )
     deepEqual( picker.get( 'select' ), null, '`select` unaffected' )
     if ( picker.get( 'min' ).pick > picker.get( 'now' ).pick ) {
         deepEqual( picker.get( 'highlight' ), picker.get( 'min' ), '`highlight` updated' )
@@ -336,7 +349,7 @@ test( '`min` using JS dates', function() {
         deepEqual( picker.get( 'highlight' ), picker.get( 'now' ), '`highlight` unaffected' )
         deepEqual( picker.get( 'view' ), picker.get( 'highlight' ), '`view` updated' )
     }
-    deepEqual( picker.get( 'max' ).pick, 1400, '`max` updated' )
+    deepEqual( picker.get( 'max' ).pick, 1410, '`max` updated' )
 })
 
 test( '`max` using integers', function() {
@@ -1050,6 +1063,7 @@ test( 'Clear', function() {
     picker.set( 'select', [2,0] )
     strictEqual( picker.get( 'value' ), picker.get( 'select', $.fn.pickatime.defaults.format ), 'Value updated' )
 
+    picker.open()
     picker.$root.find( '.' + $.fn.pickatime.defaults.klass.buttonClear ).click()
     strictEqual( picker.get( 'value' ), '', 'Value cleared' )
 })